run 2to3 and fix csv encodings
