+++ requirements.txt
@@ -1,0 +1,1 @@
+py2exe
