+++ setup.py
@@ -43,11 +43,11 @@ setup(
     name='kwaras',
     version='2.2.1.13',
     install_requires=['openpyxl'],
-    setup_requires=['py2app'],
+    setup_requires=['py2exe'],
     packages=['kwaras', 'kwaras.langs', 'kwaras.conf', 'kwaras.formats', 'kwaras.process'],
     package_dir={'kwaras': '',
-                 'kwaras.langs': 'src/langs', 'kwaras.conf': 'src/conf',
-                 'kwaras.formats': 'src/formats', 'kwaras.process': 'src/process'},
+                 'kwaras.langs': 'kwaras/langs', 'kwaras.conf': 'kwaras/conf',
+                 'kwaras.formats': 'kwaras/formats', 'kwaras.process': 'kwaras/process'},
     url='http://github.com/ucsd-field-lab/kwaras',
     license='MIT-LICENSE',
     author='Lucien Carroll',
